@@ -2,12 +2,12 @@
   import { bulletRender, bulletInitialPosition } from '$stores/coordsStores'
   import Bullet from './Bullet/Bullet.svelte'
 
-  $: bulletPairs = $bulletRender.map((bullet, index) => {
+  let bulletPairs = $derived($bulletRender.map((bullet, index) => {
     console.log('bullet', bullet)
     const initialBullet = $bulletInitialPosition.find((b) => b.id === bullet.id)
     const uniqueKey = `${bullet.id}-${index}-${initialBullet?.coords.x}-${initialBullet?.coords.y}`
     return { bullet, initialBullet, uniqueKey }
-  })
+  }))
 </script>
 
 {#each bulletPairs as { bullet, initialBullet, uniqueKey } (uniqueKey)}
