@@ -2,7 +2,7 @@
   import { T } from '@threlte/core'
   import { OrbitControls } from '@threlte/extras'
   import { useThrelte } from '@threlte/core'
-  export let camera
+  let { camera = $bindable() } = $props();
 
   let {renderer} = useThrelte()
 </script>
