@@ -2,7 +2,7 @@
   import { isEnded } from '$stores/gameStores'
   import { onDestroy } from 'svelte'
   import { goto } from '$app/navigation'
-  let win = false
+  let win = $state(false)
 
   const unsubscribe = isEnded.subscribe((value) => {
     win = value
@@ -23,7 +23,7 @@
       <h2 class="text-2xl font-semibold mb-4">You Win</h2>
       <button
         class="mt-4 px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600 transition"
-        on:click={() => {
+        onclick={() => {
           goto('/games')
         }}
       >
