@@ -1,4 +1,6 @@
 <script lang="ts">
+  import { run } from 'svelte/legacy';
+
   import {
     characterIds,
     currentPlayerId,
@@ -7,12 +9,14 @@
   import Button from '$lib/ui/Button.svelte'
   import Banner from './Banner.svelte'
 
-  let isWinner: boolean | null = null
-  $: if ($currentPlayerId != null) {
-    // As defined in the contract, the winner is the player whose character id is not 0
-    // id 0 means character is dead
-    isWinner = !($characterIds[$currentPlayerId] === 0)
-  }
+  let isWinner: boolean | null = $state(null)
+  run(() => {
+    if ($currentPlayerId != null) {
+      // As defined in the contract, the winner is the player whose character id is not 0
+      // id 0 means character is dead
+      isWinner = !($characterIds[$currentPlayerId] === 0)
+    }
+  });
 </script>
 
 <Banner color={isWinner ? '#facc15' : '#2563eb'}>
