@@ -23,7 +23,7 @@
           class={cn('rounded-md p-1 border-2', {
             'bg-white': $birdView,
           })}
-          on:click={() => ($birdView = true)}
+          onclick={() => ($birdView = true)}
         >
           <Camera color={$birdView ? 'black' : 'white'} />
         </button>
@@ -33,7 +33,7 @@
           class={cn('rounded-md p-1 border-white border-2', {
             'bg-white': !$birdView,
           })}
-          on:click={() => ($birdView = false)}
+          onclick={() => ($birdView = false)}
         >
           <User color={!$birdView ? 'black' : 'white'} />
         </button>
