@@ -4,7 +4,7 @@
   import Container from '../models/container.svelte'
   import { onMount } from 'svelte'
 
-  let container: THREE.Group = new THREE.Group()
+  let container: THREE.Group = $state(new THREE.Group())
 
   function onCreate({ ref }: any) {
     ref.traverse((object: any) => {
